Move analytics tests to the analytics module
COSDK-500
