+++ core/src/test/java/com/adyen/checkout/core/analytics/internal/data/remote/AnalyticsTrackRequestProviderTest.kt
@@ -1,8 +1,15 @@
-package com.adyen.checkout.core.internal.analytics.data.remote
+/*
+ * Copyright (c) 2025 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by ararat on 27/6/2025.
+ */
+
+package com.adyen.checkout.core.analytics.internal.data.remote
 
 import com.adyen.checkout.core.analytics.internal.AnalyticsEvent
 import com.adyen.checkout.core.analytics.internal.DirectAnalyticsEventCreation
-import com.adyen.checkout.core.analytics.internal.data.remote.AnalyticsTrackRequestProvider
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackError
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackInfo
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackLog
