+++ core/src/test/java/com/adyen/checkout/core/analytics/internal/data/remote/DefaultAnalyticsSetupProviderTest.kt
@@ -1,4 +1,12 @@
-package com.adyen.checkout.core.internal.analytics.data.remote
+/*
+ * Copyright (c) 2025 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by ararat on 27/6/2025.
+ */
+
+package com.adyen.checkout.core.analytics.internal.data.remote
 
 import android.app.Application
 import android.content.res.Resources
@@ -9,7 +17,6 @@ import com.adyen.checkout.core.analytics.internal.AnalyticsPlatformParams
 import com.adyen.checkout.core.analytics.internal.AnalyticsSource
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupRequest
 import com.adyen.checkout.core.analytics.internal.AnalyticsParamsLevel
-import com.adyen.checkout.core.analytics.internal.data.remote.DefaultAnalyticsSetupProvider
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.doReturn
